Updated model. Added error handling

+++ FinancialThing.Models/TotalAssets.cs
@@ -6,6 +6,6 @@ namespace FinancialThing.Models
     {
         public virtual int Year { get; set; }
         public virtual Financials Financials { get; set; }
-        public virtual float Value { get; set; }
+        public virtual decimal Value { get; set; }
     }
 }
